deploy: enable podInfoOnMount flag in csidriver object
This flag enables CSI driver to get the pod name, namespace, uid
and service account as part of the nodepublish request.
This come handy while we enhance features like encryption, namespace
transfer...etc. Enabling this flag as part of this patchset. This is
similar to the extra metadata CSI driver receive as part of PVC but
for pod.

Fixes https://github.com/ceph/ceph-csi/issues/2441

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>