rbd: retry creating snapshot in case flattening is in progress
When `RESOURCE_EXHAUSTED` is returned, Kubernetes aborts the
CreateSnapshot request, and does not retry it. It is expected that
flattening was started for some RBD images, so a next call to
CreateSnapshot might succeed.

Instead of returning `RESOURCE_EXHAUSTED`, return `ABORTED`, so that
Kubernetes retries CreateSnapshot with an exponential backoff.

Fixes: #2469
Signed-off-by: Niels de Vos <ndevos@redhat.com>