cephfs: save space in the cluster by avoiding roundoff
CephFS controller round off ( by util.RoundOffBytes) on the size
passed in from the provisioner at present, however this can
cause wastage of space in the cluster. The CephFS filesystem is
capable of creating volume with the exact size passed in. so no
need of Roundoff here before we reach out to the cluster

```
$ ceph fs subvolume create <vol_name> <subvol_name> [--size
<size_in_bytes>]...
```

Clone scenario:

This shouldnt be an issue as cephfs keep the parent size while the
clone is performed, we also have an extra measure at clone to make
the requested size is achieved by resize if the backend is not
having the requested size. This shouldnt be a case falling into with
this patch though.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>