cephfs: Fix umounting not working when the client has been evicted
If the CephFS client is evicted, ex: due to the node being disconnected
from the network for more than 300 seconds[1], the kernel will return
"Permission Denied" when trying to access the mount[2]. In that case we
should just ignore the error and unmount anyways.

This was tested with the kernel's CephFS driver (4.19.202).

[1] https://docs.ceph.com/en/latest/cephfs/eviction/#automatic-client-eviction
[1] https://github.com/torvalds/linux/blob/26291c54e111ff6ba87a164d85d4a4e134b7315c/fs/ceph/mds_client.c#L2809-L2820

Fix #970

Signed-off-by: Kristian Klausen <kristian@klausen.dk>